Fix remaining pull request feedback
