+++ eslint.config.mjs
@@ -24,10 +24,14 @@ export default [
       '**/.generated/**',
       'build',
       '.svelte-kit',
+      '.vercel',
       '**/.svelte-kit',
+      '**/.vercel/**',
       '**/.turbo',
       '**/build/**',
       '**/.svelte-kit/**',
+      '**/test-results/**',
+      '**/playwright-report/**',
       '**/package/**',
       'package',
       'build',
