+++ packages/scripts/content-repository.ts
@@ -1,2 +1,2 @@
-export { collectContentRepository } from './content-repository-core.ts';
-export type { ContentRepository } from './content-repository-core.ts';
+export { collectContentRepository } from './content-repository/collect.ts';
+export type { ContentRepository } from './content-repository/types.ts';
