+++ packages/scripts/content-repository/constants.ts
@@ -1,0 +1,15 @@
+export const writingReservedSlugs = new Set(['page', 'rss', 'open-graph.jpg', 'llms.txt']);
+export const courseReservedSlugs = new Set(['open-graph.jpg', 'llms.txt']);
+export const lessonReservedSlugs = new Set(['open-graph.jpg', 'llms.txt']);
+export const staticRoutes = new Set([
+  '/',
+  '/courses',
+  '/courses/open-graph.jpg',
+  '/llms-full.txt',
+  '/llms.txt',
+  '/open-graph.jpg',
+  '/sitemap.xml',
+  '/writing',
+  '/writing/open-graph.jpg',
+  '/writing/rss',
+]);
