+++ turbo.json
@@ -9,6 +9,7 @@
         "content-build.ts",
         "content-paths.ts",
         "content-repository.ts",
+        "content-repository/**/*.ts",
         "sync-generated-browser-assets.ts",
         "../utilities/package.json",
         "../utilities/content-types.ts",
@@ -40,6 +41,7 @@
         "content-build.ts",
         "content-paths.ts",
         "content-repository.ts",
+        "content-repository/**/*.ts",
         "../utilities/package.json",
         "../utilities/content-types.ts",
         "../utilities/frontmatter.ts",
@@ -77,6 +79,7 @@
         "content-build.ts",
         "content-paths.ts",
         "content-repository.ts",
+        "content-repository/**/*.ts",
         "../utilities/content-types.ts",
         "../utilities/frontmatter.ts",
         "../utilities/tailwind-playground.ts",
